Remove runCMake logic from plugin
